+++ Holovibes/includes/gui/windows/panels/import_panel.hh
@@ -55,23 +55,25 @@ class ImportPanel : public Panel
      */
     void update_fps();
 
-    /**
-     * @brief Handles the update of the import file path in the UI.
+    /*!
+     * \brief Handles the update of the import file path in the UI.
      */
     void update_import_file_path();
 
-    /**
-     * @brief Handles the update of the load file in GPU in the UI.
+    /*!
+     * \brief Handles the update of the load file kind in the UI.
+     *
+     * \param kind[in] The index of the new way to read from files
      */
-    void update_load_file_in_gpu();
+    void update_file_load_kind(int kind);
 
-    /**
-     * @brief Handles the update of the input file start index in the UI.
+    /*!
+     * \brief Handles the update of the input file start index in the UI.
      */
     void update_input_file_start_index();
 
-    /**
-     * @brief Handles the update of the input file end index in the UI.
+    /*!
+     * \brief Handles the update of the input file end index in the UI.
      */
     void update_input_file_end_index();
 };