+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - time_transformation_size from 16 to 32
   - time_transformation_stride 32 to 64
   - z width from 8 to 16
+- Changed the UI option 'Load file in GPU' to a drop-down menu and renamed it to 'Load file in GPU VRAM'
+- Added an option 'Load file in CPU RAM' in that drop-down
 
 ### 14.2.0